backend/requirements.txt
annotated-types
anyio
APScheduler
argon2-cffi
argon2-cffi-bindings
bcrypt
CacheControl
cachetools
certifi
cffi
charset-normalizer
click
cryptography
dnspython
email-validator
exceptiongroup
fastapi
fastapi-cli
firebase-admin
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
orjson==3.11.3
passlib==1.7.4
proto-plus==1.26.1
protobuf==6.32.1
# psycopg2-binary retiré - utilisation de SQLite locale maintenant
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
#pydantic==1.10.15
pydantic_core==2.18.4
Pygments==2.19.2
PyJWT==2.8.0
pyparsing==3.2.4
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.17.4
typing_extensions==4.15.0
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
